+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning][].
 
 ### Changed
 
+* `a2s` query methods accept `context.Context` for cancellation
+  and total query deadlines
 * `a2s` player query methods return slices directly
   instead of pointers to slices
 * `a2s` expose raw and parsed A2S_RULES views